SPEED-777: Remove all Railgun-related code
**🔖 Summary**

For the last few weeks, the users of the Cloudflare Wordpress plugin
are seeing `Bad Request` errors in the form of a banner on the
Cloudflare plugin interface. The bad requests are actually
Railgun-relatde calls performed by this Javascript dependency.

Since Railgun is now deprecated, this commit takes the approach of
removing all Railgun-related code to make sure no calls are made against
the Cloudflare API to retrieve the status of the Railgun configuration.

**✅ Testing plan**

- All existing tests are expected to pass
- Next published version of the Wordpress plugin should not display any
  Bad Request banner related to Railgun failing requests